Fix - move ASP.NET Core request tracking to seperate library
