+++ src/Arcus.Observability.Tests.Unit/Arcus.Observability.Tests.Unit.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Arcus.Observability.Correlation\Arcus.Observability.Correlation.csproj" />
+    <ProjectReference Include="..\Arcus.Observability.Telemetry.AspNetCore\Arcus.Observability.Telemetry.AspNetCore.csproj" />
     <ProjectReference Include="..\Arcus.Observability.Telemetry.IoT\Arcus.Observability.Telemetry.IoT.csproj" />
     <ProjectReference Include="..\Arcus.Observability.Telemetry.AzureFunctions\Arcus.Observability.Telemetry.AzureFunctions.csproj" />
     <ProjectReference Include="..\Arcus.Observability.Telemetry.Serilog.Filters\Arcus.Observability.Telemetry.Serilog.Filters.csproj" />
